+++ app/components/science/index.ts
@@ -1,6 +1,8 @@
 export { default as ScienceHero } from "./ScienceHero";
 export { default as TwoSystemModel } from "./TwoSystemModel";
 export { default as ScienceAdaptogens } from "./ScienceAdaptogens";
+export { default as ScienceNootropics } from "./ScienceNootropics";
+export { default as ScienceExplainer } from "./ScienceExplainer";
 export { default as SciencePillars } from "./SciencePillars";
 export { default as PillarCard } from "./PillarCard";
 export { default as SynergyChart } from "./SynergyChart";